+++ amulet/level/formats/anvil_world/__init__.py
@@ -1,0 +1,4 @@
+# meta format
+from .format import AnvilFormat
+
+export = AnvilFormat
