+++ amulet/level/formats/leveldb_world/__init__.py
@@ -1,0 +1,4 @@
+# meta format
+from .format import LevelDBFormat
+
+export = LevelDBFormat
