--- amulet/world_interface/formats/anvil/format.meta
@@ -1,10 +1,0 @@
-{
-"meta_version": 0,
-"format": {
-    "id": "java",
-    "name": "Anvil Format",
-    "entry_point": "anvil_format",
-    "format_version": 0,
-    "wrapper_version": -1
-   }
-}
