--- amulet/world_interface/formats/leveldb/format.meta
@@ -1,10 +1,0 @@
-{
-"meta_version": 0,
-"format": {
-    "id": "bedrock",
-    "name": "Leveldb Format",
-    "entry_point": "leveldb_format",
-    "format_version": 0,
-    "wrapper_version": -1
-   }
-}
